updated to store received event only
