+++ src/main/java/uk/gov/ons/census/casesvc/service/RefusalProcessor.java
@@ -47,7 +47,6 @@ public class RefusalProcessor {
     }
 
     Case caze = optCase.get();
-
     caze.setRefusalReceived(true);
     caseRepository.saveAndFlush(caze);
