+++ src/test/java/uk/gov/ons/census/casesvc/service/QuestionnaireLinkedProcessorTest.java
@@ -6,6 +6,7 @@ import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 import static uk.gov.ons.census.casesvc.testutil.DataUtils.getRandomCase;
 import static uk.gov.ons.census.casesvc.testutil.DataUtils.getTestResponseManagementQuestionnaireLinkedEvent;
+import static uk.gov.ons.census.casesvc.utility.JsonHelper.convertObjectToJson;
 
 import java.util.Optional;
 import java.util.UUID;
@@ -57,11 +58,11 @@ public class QuestionnaireLinkedProcessorTest {
     // then
     verify(uacProcessor, times(1)).emitUacUpdatedEvent(uacQidLink, caze);
     verify(eventLogger, times(1))
-        .logQuestionnaireLinkedEvent(
+        .logEvent(
             uacQidLink,
             "Questionnaire Linked",
             EventType.QUESTIONNAIRE_LINKED,
-            uac,
+            convertObjectToJson(uac),
             managementEvent.getEvent());
   }
